[MISC] move include/seqan3/{core,utility}/simd/detail/builtin_simd_intrinsics.hpp
