+++ include/seqan3/core/simd/detail/simd_algorithm_sse4.hpp
@@ -15,7 +15,7 @@
 #include <array>
 
 #include <seqan3/utility/simd/concept.hpp>
-#include <seqan3/core/simd/detail/builtin_simd_intrinsics.hpp>
+#include <seqan3/utility/simd/detail/builtin_simd_intrinsics.hpp>
 #include <seqan3/core/simd/detail/builtin_simd.hpp>
 #include <seqan3/utility/simd/simd_traits.hpp>
